requirements.txt
numpy>=1.26
scipy>=1.13
scikit-learn>=1.5
pandas>=2.2
pyyaml>=6.0
matplotlib>=3.8
seaborn>=0.13
torch>=2.2; platform_system != "Windows" or platform_machine != "ARM64"
torch>=2.3; platform_system == "Windows" and platform_machine == "AMD64"
mlflow>=2.16
pdfminer.six>=20250506
einops>=0.8
tqdm>=4.66
colorlog>=6.8
pytest>=7.4

